refactor(gir): Consolidate enum-related imports and introduce new enum member class
- Updated import statements across multiple files to streamline the organization of enum-related types.
- Introduced a new `GirEnumMember` class to encapsulate enum member functionality, enhancing clarity and maintainability.
- Removed deprecated enum member functionality from the `IntrospectedEnum` class, ensuring a cleaner structure.
- Updated various files to utilize the new `GirEnumMember` class, improving consistency across the codebase.
