+++ packages/lib/src/generators/generator.ts
@@ -3,7 +3,9 @@ import type { IntrospectedClass } from "../gir/introspected-class.ts";
 import type { IntrospectedRecord } from "../gir/record.ts";
 import type { IntrospectedInterface } from "../gir/interface.ts";
 import type { IntrospectedConstant } from "../gir/const.ts";
-import type { IntrospectedEnum, IntrospectedError, GirEnumMember } from "../gir/enum.ts";
+import type { IntrospectedEnum } from "../gir/enum.ts";
+import type { IntrospectedError } from "../gir/error.ts";
+import type { GirEnumMember } from "../gir/enum-member.ts";
 import type { IntrospectedProperty, IntrospectedField } from "../gir/property.ts";
 import type { IntrospectedSignal, IntrospectedSignalType } from "../gir/signal.ts";
 import type { IntrospectedFunctionParameter } from "../gir/parameter.ts";
