+++ packages/lib/src/gir/enum.ts
@@ -1,12 +1,10 @@
 import { NumberType, TypeIdentifier } from "../gir.ts";
 
-import { IntrospectedBase } from "./introspected-base.ts";
 import { IntrospectedNamespaceMember } from "./introspected-namespace-member.ts";
-import type { GirMemberElement, GirEnumElement, GirBitfieldElement } from "../index.ts";
+import type { GirEnumElement, GirBitfieldElement } from "../index.ts";
 
 import { IntrospectedRecord } from "./record.ts";
 import { IntrospectedField } from "./property.ts";
-import { IntrospectedStaticClassFunction } from "./static-function.ts";
 import { GirComplexRecord } from "./gir-complex-record.ts";
 import { IntrospectedNamespace } from "./namespace.ts";
 
@@ -17,6 +15,7 @@ import { FormatGenerator } from "../generators/generator.ts";
 import { GirVisitor } from "../visitor.ts";
 
 import type { OptionsLoad } from "../types/index.ts";
+import { GirEnumMember } from "./enum-member.ts";
 
 export class IntrospectedEnum extends IntrospectedNamespaceMember {
     members = new Map<string, GirEnumMember>();
@@ -136,122 +135,3 @@ export class IntrospectedEnum extends IntrospectedNamespaceMember {
     }
 }
 
-export class GirEnumMember extends IntrospectedBase<IntrospectedEnum> {
-    value: string;
-    c_identifier: string;
-
-    constructor(name: string, value: string, parent: IntrospectedEnum, c_identifier: string) {
-        super(name, parent);
-        this.value = value;
-        this.c_identifier = c_identifier;
-    }
-
-    get namespace() {
-        return this.parent.namespace;
-    }
-
-    accept(visitor: GirVisitor): GirEnumMember {
-        const node = this.copy();
-        return visitor.visitEnumMember?.(node) ?? node;
-    }
-
-    copy(): GirEnumMember {
-        const { value, name, parent, c_identifier } = this;
-
-        return new GirEnumMember(name, value, parent, c_identifier)._copyBaseProperties(this);
-    }
-
-    static fromXML(element: GirMemberElement, parent: IntrospectedEnum, options: OptionsLoad): GirEnumMember {
-        const upper = element.$.name.toUpperCase();
-        const c_identifier = element.$["c:identifier"];
-
-        const enumMember = new GirEnumMember(upper, element.$.value, parent, c_identifier);
-
-        if (options.loadDocs) {
-            enumMember.doc = parseDoc(element);
-            enumMember.metadata = parseMetadata(element);
-        }
-
-        return enumMember;
-    }
-
-    asString<T extends FormatGenerator<unknown>>(generator: T): ReturnType<T["generateEnumMember"]> {
-        return generator.generateEnumMember(this) as ReturnType<T["generateEnumMember"]>;
-    }
-}
-
-function isEnumElement(e: unknown): e is GirEnumElement {
-    return typeof e === "object" && e != null && "function" in e;
-}
-
-export class IntrospectedError extends IntrospectedEnum {
-    functions: Map<string, IntrospectedStaticClassFunction> = new Map();
-
-    asString<T extends FormatGenerator<unknown>>(generator: T): ReturnType<T["generateError"]> {
-        return generator.generateError(this) as ReturnType<T["generateError"]>;
-    }
-
-    copy({
-        members
-    }: {
-        parent?: undefined;
-        members?: Map<string, GirEnumMember>;
-    } = {}): IntrospectedEnum {
-        const { namespace, name, flags } = this;
-
-        const en = new IntrospectedError(name, namespace);
-
-        for (const [key, member] of (members ?? this.members).entries()) {
-            en.members.set(key, member.copy());
-        }
-
-        for (const [key, func] of this.functions.entries()) {
-            en.functions.set(key, func.copy({ parent: en }));
-        }
-
-        en.flags = flags;
-
-        return en._copyBaseProperties(this);
-    }
-
-    static fromXML(
-        element: GirEnumElement | GirBitfieldElement,
-        ns: IntrospectedNamespace,
-        options: OptionsLoad
-    ): IntrospectedEnum {
-        const err = new IntrospectedError(sanitizeMemberName(element.$.name), ns);
-
-        if (element.$["glib:type-name"]) {
-            err.resolve_names.push(element.$["glib:type-name"]);
-
-            ns.registerResolveName(element.$["glib:type-name"], ns.namespace, err.name);
-        }
-
-        if (element.$["c:type"]) {
-            err.resolve_names.push(element.$["c:type"]);
-
-            ns.registerResolveName(element.$["c:type"], ns.namespace, err.name);
-        }
-
-        if (options.loadDocs) {
-            err.doc = parseDoc(element);
-            err.metadata = parseMetadata(element);
-        }
-
-        if (element.member) {
-            element.member.forEach(m => {
-                const member = GirEnumMember.fromXML(m, err, options);
-                err.members.set(member.name, member);
-            });
-        }
-
-        if (isEnumElement(element) && element.function) {
-            element.function.forEach(f => {
-                const func = IntrospectedStaticClassFunction.fromXML(f, err, options);
-                err.functions.set(func.name, func);
-            });
-        }
-
-        return err;
-    }
-}
