+++ packages/lib/src/gir/index.ts
@@ -6,7 +6,9 @@ export * from "./function.ts";
 export * from "./constructor.ts";
 export * from "./const.ts";
 export * from "./data.ts";
+export * from "./enum-member.ts";
 export * from "./enum.ts";
+export * from "./error.ts";
 export * from "./generics.ts";
 export * from "./gir-complex-record.ts";
 export * from "./interface.ts";
