+++ packages/lib/src/visitor.ts
@@ -5,7 +5,9 @@ import type { IntrospectedInterface } from './gir/interface.ts'
 import type { IntrospectedClass } from './gir/introspected-class.ts'
 import type { IntrospectedBaseClass } from './gir/introspected-base-class.ts'
 import type { IntrospectedConstant } from './gir/const.ts'
-import type { GirEnumMember, IntrospectedError, IntrospectedEnum } from './gir/enum.ts'
+import type { IntrospectedEnum } from './gir/enum.ts'
+import type { IntrospectedError } from './gir/error.ts'
+import type { GirEnumMember } from './gir/enum-member.ts'
 import type {
     IntrospectedFunctionParameter,
 } from './gir/parameter.ts'
